fix: remove dependency on local.properties

+++ app/src/main/kotlin/io/github/bradleyiw/ns/server/ServerUrlProvider.kt
@@ -20,7 +20,7 @@ class ServerUrlProvider @Inject constructor(
         port = port()
     )
 
-    private fun port(): Int = BuildConfig.CALL_MONITOR_SERVER_PORT.toInt()
+    private fun port(): Int = BuildConfig.CALL_MONITOR_SERVER_PORT
 
     private fun ipAddress(): String? =
         activeNetworkLinkProperties()?.let {
